+++ tests/module.test.js
@@ -24,18 +24,15 @@ test('Launch Fastify for testing', async assert => {
   assert.ok(fastifyInstance)
 
   const serverAddress = fastifyInstance.server.address()
-  assert.strictSame(serverAddress.port, 3000)
-  assert.strictSame(serverAddress.address, '127.0.0.1')
+  assert.notOk(serverAddress)
   assert.strictSame(fastifyInstance.log.level, 'silent')
 
-  await fastifyInstance.close()
   assert.end()
 })
 
 test('Launch Fastify for testing, overriding default values', async assert => {
   const fastifyInstance = await launch('./tests/modules/correct-module', {
     logLevel: 'silent',
-    port: 9000,
     envVariables: {
       ENV_KEY: 'ENV_VALUE',
     },
@@ -52,6 +49,5 @@ test('Launch Fastify for testing, overriding default values', async assert => {
     },
   })
 
-  await fastifyInstance.close()
   assert.end()
 })