+++ CHANGELOG.md
@@ -1,10 +1,17 @@
 # Changes to `lc39`
 
+### TBR
+
+- The exported module will create a `fastify` instace that
+  is not listening on any port.
+
 ### v1.1.0 (2019-04-10)
+
 - Add `errorHandler` options for customize `fastify-sensible`
   auto handling of uncatched errors
 
 ### v1.0.0 (2019-04-09)
+
 - Add `fastify-sensible` plugin integration
 
 ### v0.4.0 (2019-04-08)
@@ -22,12 +29,15 @@
 - Use the `fastify` embedded version of `pino`
 
 ### v0.3.0 (2019-02-06)
+
 - Fixed `ENV` injection in `testLaunch` function
 - Fixed override of default values in `testLaunch` function
 - Relincesed under Apache 2.0 from MIT
 
 ### v0.2.0 (2019-01-16)
+
 - Add `/-/healthz and` `/-/ready` status endpoints.
 
 ### v0.1.0 (2019-01-08)
+
 - Initial relase